allow users to override source of npmrc file via hiera.
rather than having an ensure line to modify the npmrc file, it would
be easier to manage the whole file directly.  We accomplish this by
moving the source of the npmrc file to a parameter for the class,
currently defaulted to the previously hardcoded source file.  In heira,
this can be overridden with any file the user likes.

WARNING: this does allow users to not set artifactory as their default
npm registry.  Perhaps we should have an ensure line to make sure the
template includes this?