update profile
having both dollar signs when doing string interpolation is very unnecessary and potentially not correct.  updating the fish class to reference dotFiles to get the correct location where the dotfiles should go.  This will help consolidate logic and pre-emptively fix a bug if the dotfile directory were changed to be modifiable via hiera.